packages/workers/pyproject.toml
name = "workers"
version = "0.1.0"
description = "Python video processing workers for MusicApp"

[tool.poetry.dependencies]
python = "^3.11"
redis = "^5.0.1"
librosa = "^0.10.0"
numpy = "^1.24.0"
opencv-python = "^4.8.0"
torch = "^2.1.0"
torchaudio = "^2.1.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
pydantic = "^2.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
